refactor(cn-efs): Remove unccessary properties from tailwindOrWindiCN_EFS filter object
I decided to remove all the properties that were associated with the conflict between a single word class and one with a value.
I also decided to chage the type of the filter object so that I dont need a lowercase word. I felt like having to write quotes is ugly.