refactor(class-map-changers):change SortedClassses to ClassNamesMap.
I decided to change all the names from sorted classes to classNames map because it's correct.
I also decided to use the term utility because it's the right term.
I also decided to remove the type that unified all the class name maps together.
It was broken by the unified utility change.